requirements.txt
gunicorn
numpy
pandas
PyYAML
matplotlib
Flask
sklearn
-e .